@@ -4,6 +4,7 @@ plugins {
     id("com.android.application")
     id("org.jetbrains.kotlin.android")
     id("org.jlleitschuh.gradle.ktlint")
+    id("kotlin-parcelize")
     id("kotlin-kapt")
     id("androidx.navigation.safeargs.kotlin")
 }
@@ -24,8 +25,20 @@ android {
     }
 
     compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_1_8
-        targetCompatibility = JavaVersion.VERSION_1_8
+        sourceCompatibility = JavaVersion.VERSION_17
+        targetCompatibility = JavaVersion.VERSION_17
+    }
+
+    kotlinOptions {
+        jvmTarget = "17"
+        freeCompilerArgs = freeCompilerArgs + listOf(
+            "-opt-in=kotlin.RequiresOptIn",
+            "-Xjvm-default=enable",
+        )
+    }
+
+    buildFeatures {
+        viewBinding = true
     }
 
     buildTypes {
@@ -39,37 +52,29 @@ android {
 
 dependencies {
     implementation(project(":transitionx"))
-
-    implementation("androidx.core:core-ktx:1.0.0")
-
+    implementation("androidx.core:core-ktx:1.10.1")
     implementation("org.jetbrains.kotlin:kotlin-stdlib:1.8.21")
 
     // Support
-    implementation("com.google.android.material:material:1.0.0")
     implementation("androidx.cardview:cardview:1.0.0")
-    implementation("androidx.vectordrawable:vectordrawable:1.0.0")
-    implementation("androidx.recyclerview:recyclerview:1.0.0")
+    implementation("androidx.vectordrawable:vectordrawable:1.1.0")
+    implementation("androidx.recyclerview:recyclerview:1.3.0")
+    implementation("androidx.constraintlayout:constraintlayout:2.1.4")
 
-    implementation("androidx.constraintlayout:constraintlayout:1.1.3")
+    implementation("com.google.android.material:material:1.9.0")
+    implementation("com.google.android.flexbox:flexbox:3.0.0")
 
-    implementation("com.google.android.flexbox:flexbox:${Versions.flexBox}")
-
-    // Logging
-    implementation("com.jakewharton.timber:timber:${Versions.timber}")
-
-    // Glide - Image loading
-    implementation("com.github.bumptech.glide:glide:${Versions.glide}")
-    kapt("com.github.bumptech.glide:compiler:${Versions.glide}")
+    implementation("com.hannesdorfmann:adapterdelegates3:3.0.1")
+    implementation("com.jakewharton.timber:timber:5.0.1")
+    implementation("com.github.bumptech.glide:glide:4.13.2")
+    kapt("com.github.bumptech.glide:compiler:4.13.2")
 
     // Navigation
-    implementation("androidx.navigation:navigation-fragment:2.0.0-rc02")
-    implementation("androidx.navigation:navigation-ui:2.0.0-rc02")
-    implementation("androidx.navigation:navigation-ui:2.0.0-rc02")
-    implementation("androidx.navigation:navigation-ui-ktx:2.0.0-rc02")
+    implementation("androidx.navigation:navigation-fragment:2.6.0")
+    implementation("androidx.navigation:navigation-ui:2.6.0")
+    implementation("androidx.navigation:navigation-ui:2.6.0")
+    implementation("androidx.navigation:navigation-ui-ktx:2.6.0")
 
-    // Adapter Delegates
-    implementation("com.hannesdorfmann:adapterdelegates3:${Versions.adapterDelegates}")
-
-    testImplementation("junit:junit:${Versions.jUnit}")
-    testImplementation("org.robolectric:robolectric:${Versions.roboElectric}")
+    // Testing
+    testImplementation("junit:junit:4.13.2")
 }
