@@ -1,34 +1,18 @@
-/*
- *
- * Copyright 2019 Arunkumar
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
 package `in`.arunkumarsampath.transitionx.sample.home
 
-import `in`.arunkumarsampath.transitionx.prepareTransition
-import `in`.arunkumarsampath.transitionx.sample.R
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.isGone
 import androidx.navigation.NavController
 import androidx.navigation.findNavController
 import androidx.navigation.ui.NavigationUI
-import kotlinx.android.synthetic.main.activity_main.*
+import `in`.arunkumarsampath.transitionx.prepareTransition
+import `in`.arunkumarsampath.transitionx.sample.R
+import `in`.arunkumarsampath.transitionx.sample.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
+
+    private lateinit var binding: ActivityMainBinding
 
     private val navController: NavController by lazy {
         findNavController(R.id.nav_host_fragment)
@@ -36,18 +20,20 @@ class MainActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
         setupBottomNavigationMenu()
     }
 
     private fun setupBottomNavigationMenu() {
-        NavigationUI.setupWithNavController(bottomNavigation, navController)
+        NavigationUI.setupWithNavController(binding.bottomNavigation, navController)
         navController.addOnDestinationChangedListener { _, destination, _ ->
-            rootCoordinatorLayout.prepareTransition {
-                +bottomNavigation
+            binding.rootCoordinatorLayout.prepareTransition {
+                +binding.bottomNavigation
                 slide()
             }
-            bottomNavigation.isGone = destination.id != R.id.homeFragment
+            binding.bottomNavigation.isGone = destination.id != R.id.homeFragment
         }
     }
 
