@@ -5,23 +5,22 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.LinearLayout
-import androidx.core.app.Fragment
 import androidx.core.view.isGone
 import androidx.core.view.updateLayoutParams
+import androidx.fragment.app.Fragment
 import `in`.arunkumarsampath.transitionx.prepareTransition
-import `in`.arunkumarsampath.transitionx.sample.R
-import `in`.arunkumarsampath.transitionx.sample.extensions.dpToPx
-import kotlinx.android.synthetic.main.layout_explode_transition_content.*
+import `in`.arunkumarsampath.transitionx.sample.databinding.FragmentExplodeTransitionBinding
+import `in`.arunkumarsampath.transitionx.sample.utils.dpToPx
 
 class ExplodeTransitionFragment : Fragment() {
 
+    private lateinit var binding: FragmentExplodeTransitionBinding
     private var toggle = true
 
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?,
-    ): View? = inflater.inflate(R.layout.fragment_explode_transition, container, false)
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+        binding = FragmentExplodeTransitionBinding.inflate(inflater, container, false)
+        return binding.root
+    }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -29,34 +28,34 @@ class ExplodeTransitionFragment : Fragment() {
     }
 
     private fun initViewListeners() {
-        fab.setOnClickListener {
-            frameLayout.prepareTransition {
+        binding.layoutContent.fab.setOnClickListener {
+            binding.layoutContent.frameLayout.prepareTransition {
                 explode {
-                    +accentBackground
+                    +binding.layoutContent.accentBackground
                 }
                 moveResize {
-                    +userIconView
+                    +binding.layoutContent.userIconView
                 }
             }
 
             if (toggle) {
                 with(requireContext()) {
-                    userIconView.updateLayoutParams<LinearLayout.LayoutParams> {
+                    binding.layoutContent.userIconView.updateLayoutParams<LinearLayout.LayoutParams> {
                         height = dpToPx(112.0)
                         width = dpToPx(112.0)
                     }
                 }
-                accentBackground.isGone = true
-                helloText.isGone = false
+                binding.layoutContent.accentBackground.isGone = true
+                binding.layoutContent.helloText.isGone = false
             } else {
                 with(requireContext()) {
-                    userIconView.updateLayoutParams<LinearLayout.LayoutParams> {
+                    binding.layoutContent.userIconView.updateLayoutParams<LinearLayout.LayoutParams> {
                         height = dpToPx(56.0)
                         width = dpToPx(56.0)
                     }
                 }
-                accentBackground.isGone = false
-                helloText.isGone = true
+                binding.layoutContent.accentBackground.isGone = false
+                binding.layoutContent.helloText.isGone = true
             }
             toggle = !toggle
         }
