@@ -8,6 +8,7 @@ import android.graphics.drawable.GradientDrawable
 import android.os.Bundle
 import android.view.View
 import android.view.ViewGroup
+import android.widget.ImageView
 import android.widget.TextView
 import androidx.annotation.ColorInt
 import androidx.annotation.ColorRes
@@ -19,13 +20,14 @@ import androidx.transition.TransitionValues
 import `in`.arunkumarsampath.transitionx.evaluator.ArgbEvaluator
 import `in`.arunkumarsampath.transitionx.prepareTransition
 import `in`.arunkumarsampath.transitionx.sample.R
-import `in`.arunkumarsampath.transitionx.sample.extensions.resolveColorRes
+import `in`.arunkumarsampath.transitionx.sample.databinding.LayoutAnimatedBottomNavigationContentBinding
+import `in`.arunkumarsampath.transitionx.sample.utils.resolveColorRes
 import `in`.arunkumarsampath.transitionx.transitionSet
-import kotlinx.android.synthetic.main.layout_animated_bottom_navigation_content.*
 import kotlin.LazyThreadSafetyMode.NONE
 
 class AnimatedBottomNavigationFragment : BaseSampleFragment() {
 
+    private lateinit var binding: LayoutAnimatedBottomNavigationContentBinding
     override val contentLayoutResource = R.layout.layout_animated_bottom_navigation_content
     override val titleRes = R.string.animated_bottom_navigation_fragment
 
@@ -33,7 +35,7 @@ class AnimatedBottomNavigationFragment : BaseSampleFragment() {
      * Data structure for holding components in each bottom nav item.
      */
     private data class BottomNavigationItem(
-        val icon: AppCompatImageView,
+        val icon: ImageView,
         val text: TextView,
         val bg: View,
         @param:ColorRes val colorRes: Int,
@@ -44,15 +46,16 @@ class AnimatedBottomNavigationFragment : BaseSampleFragment() {
      */
     private val navItems by lazy(NONE) {
         sequenceOf(
-            BottomNavigationItem(home, homeText, homeBg, R.color.bottom_nav_home),
-            BottomNavigationItem(likes, likesText, likesBg, R.color.bottom_nav_likes),
-            BottomNavigationItem(search, searchText, searchBg, R.color.bottom_nav_search),
-            BottomNavigationItem(about, aboutText, aboutBg, R.color.bottom_nav_about),
+            BottomNavigationItem(binding.home, binding.homeText, binding.homeBg, R.color.bottom_nav_home),
+            BottomNavigationItem(binding.likes, binding.likesText, binding.likesBg, R.color.bottom_nav_likes),
+            BottomNavigationItem(binding.search, binding.searchText, binding.searchBg, R.color.bottom_nav_search),
+            BottomNavigationItem(binding.about, binding.aboutText, binding.aboutBg, R.color.bottom_nav_about),
         )
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        binding = LayoutAnimatedBottomNavigationContentBinding.bind(viewStubRoot)
         bindListeners()
     }
 
@@ -61,7 +64,7 @@ class AnimatedBottomNavigationFragment : BaseSampleFragment() {
      * but no sequentially. Instead it relies on custom [Transition.setStartDelay] to achieve desired
      * choreography.
      *
-     * Additionally, it also animates changes to [AppCompatImageView.setImageTint] property.
+     * Additionally, it also animates changes to [ImageView.setImageTint] property.
      */
     private val bottomNavTransition by lazy {
         transitionSet {
@@ -80,7 +83,7 @@ class AnimatedBottomNavigationFragment : BaseSampleFragment() {
 
             changeColor {
                 navItems.map { it.text }.forEach { text -> add(text) }
-                +constraintLayout
+                +binding.constraintLayout
             }
 
             customTransition<ChangeImageTint> {
@@ -93,7 +96,7 @@ class AnimatedBottomNavigationFragment : BaseSampleFragment() {
         navItems.forEach { navItem ->
             navItem.icon.setOnClickListener {
                 // Prepare transition
-                constraintLayout.prepareTransition(cancel = true, transition = bottomNavTransition)
+                binding.constraintLayout.prepareTransition(cancel = true, transition = bottomNavTransition)
 
                 // Reset all other items
                 navItems.filter { it != navItem }
@@ -122,19 +125,19 @@ class AnimatedBottomNavigationFragment : BaseSampleFragment() {
                         (background.mutate() as GradientDrawable).setColor(setAlphaComponent(selectedColor, (0.25 * 255).toInt()))
                     }
 
-                    constraintLayout.background = ColorDrawable(selectedColor)
+                    binding.constraintLayout.background = ColorDrawable(selectedColor)
                 }
             }
         }
     }
 
     /**
-     * Animates changes to [AppCompatImageView.setImageTint]
+     * Animates changes to [ImageView.setImageTint]
      */
     private class ChangeImageTint : Transition() {
 
         private fun captureValues(transitionValues: TransitionValues) {
-            val imageView = transitionValues.view as? AppCompatImageView
+            val imageView = transitionValues.view as? ImageView
             imageView?.let {
                 val color = ImageViewCompat.getImageTintList(imageView)?.defaultColor ?: -1
                 transitionValues.values[PROPNAME_IMAGE_TINT] = color
@@ -156,7 +159,7 @@ class AnimatedBottomNavigationFragment : BaseSampleFragment() {
                 return null
             }
 
-            val imageView = endValues.view as? AppCompatImageView
+            val imageView = endValues.view as? ImageView
             val startTint = startValues.values[PROPNAME_IMAGE_TINT] as Int
             val endTint = endValues.values[PROPNAME_IMAGE_TINT] as Int
 
@@ -184,11 +187,11 @@ class AnimatedBottomNavigationFragment : BaseSampleFragment() {
 
     companion object {
 
-        private fun AppCompatImageView.setImageTintColor(@ColorInt colorInt: Int) {
+        private fun ImageView.setImageTintColor(@ColorInt colorInt: Int) {
             ImageViewCompat.setImageTintList(this, ColorStateList.valueOf(colorInt))
         }
 
-        private fun AppCompatImageView.setImageTint(@ColorRes colorRes: Int) {
+        private fun ImageView.setImageTint(@ColorRes colorRes: Int) {
             ImageViewCompat.setImageTintList(
                 this,
                 ColorStateList.valueOf(context.resolveColorRes(colorRes)),
