@@ -2,19 +2,19 @@ package `in`.arunkumarsampath.transitionx.sample.home.transitionsamples
 
 import android.os.Bundle
 import android.view.View
-import androidx.constraintlayout.ConstraintSet
+import androidx.constraintlayout.widget.ConstraintSet
 import androidx.core.view.isGone
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.drawable.DrawableTransitionOptions.withCrossFade
 import com.bumptech.glide.request.RequestOptions
 import `in`.arunkumarsampath.transitionx.prepareTransition
 import `in`.arunkumarsampath.transitionx.sample.R
-import `in`.arunkumarsampath.transitionx.sample.extensions.removeCallbacks
+import `in`.arunkumarsampath.transitionx.sample.databinding.LayoutMaterialCardTransformationBinding
 import `in`.arunkumarsampath.transitionx.transition.changetext.ChangeText
-import kotlinx.android.synthetic.main.layout_material_card_transformation.*
 
 class MaterialCardTransformationFragment : BaseSampleFragment() {
 
+    private lateinit var binding: LayoutMaterialCardTransformationBinding
     override val contentLayoutResource = R.layout.layout_material_card_transformation
     override val titleRes = R.string.metamorphosis
 
@@ -34,10 +34,11 @@ class MaterialCardTransformationFragment : BaseSampleFragment() {
     private val expandRunnable = { expandTransition() }
     private val collapseRunnable = { collapseTransition() }
 
-    private val imageViews = arrayOf(image1, image2, image3, image4, image5)
+    private val imageViews by lazy { arrayOf(binding.image1, binding.image2, binding.image3, binding.image4, binding.image5) }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        binding = LayoutMaterialCardTransformationBinding.bind(viewStubRoot)
         setupViews()
     }
 
@@ -46,22 +47,23 @@ class MaterialCardTransformationFragment : BaseSampleFragment() {
             .load(R.drawable.ic_girl)
             .transition(withCrossFade())
             .apply(RequestOptions().circleCrop())
-            .into(avatar)
+            .into(binding.avatar)
         listOf(
-            R.drawable.ic_artic_1 to image1,
-            R.drawable.ic_artic_2 to image2,
-            R.drawable.ic_artic_3 to image3,
-            R.drawable.ic_artic_4 to image4,
-            R.drawable.ic_artic_5 to image5,
+            R.drawable.ic_artic_1 to binding.image1,
+            R.drawable.ic_artic_2 to binding.image2,
+            R.drawable.ic_artic_3 to binding.image3,
+            R.drawable.ic_artic_4 to binding.image4,
+            R.drawable.ic_artic_5 to binding.image5,
         ).forEach { (drawable, image) ->
             requestManager.load(drawable)
                 .transition(withCrossFade())
                 .apply(RequestOptions().centerCrop())
                 .into(image)
         }
-        with(collapseButton) {
+        with(binding.collapseButton) {
             setOnClickListener {
-                removeCallbacks(expandRunnable, collapseRunnable)
+                // removeCallbacks(expandRunnable, collapseRunnable)
+                removeCallbacks(expandRunnable)
                 val runnable = if (expand) expandRunnable else collapseRunnable
                 postDelayed(runnable, 100) // Add delay to allow ripple animation to run
                 expand = !expand
@@ -70,62 +72,62 @@ class MaterialCardTransformationFragment : BaseSampleFragment() {
     }
 
     private fun collapseTransition() {
-        constraintLayout.prepareTransition {
+        binding.constraintLayout.prepareTransition {
             auto {
                 ease { standardEasing }
-                exclude(metamorphosisDesc2)
+                exclude(binding.metamorphosisDesc2)
             }
             transitionSet {
                 fade()
                 slide()
                 ease { accelerateEasing }
-                +metamorphosisDesc2
+                +binding.metamorphosisDesc2
             }
             changeImage { add(*imageViews) }
             onEnd {
-                constraintLayout.prepareTransition {
+                binding.constraintLayout.prepareTransition {
                     moveResize()
                     changeText {
-                        +collapseButton
+                        +binding.collapseButton
                         changeTextBehavior = ChangeText.CHANGE_BEHAVIOR_OUT_IN
                     }
                 }
-                collapseButton.setText(R.string.expand)
+                binding.collapseButton.setText(R.string.expand)
             }
             duration = 300
         }
-        collapseConstraint.applyTo(constraintLayout)
-        metamorphosisDesc2.isGone = true
-        metamorphosisDesc.isGone = false
+        collapseConstraint.applyTo(binding.constraintLayout)
+        binding.metamorphosisDesc2.isGone = true
+        binding.metamorphosisDesc.isGone = false
     }
 
     private fun expandTransition() {
-        constraintLayout.prepareTransition {
+        binding.constraintLayout.prepareTransition {
             auto {
                 ease { standardEasing }
-                exclude(metamorphosisDesc2)
+                exclude(binding.metamorphosisDesc2)
             }
             transitionSet {
                 fade()
                 slide()
                 ease { accelerateEasing }
-                +metamorphosisDesc2
+                +binding.metamorphosisDesc2
             }
             changeImage { add(*imageViews) }
             onEnd {
-                constraintLayout.prepareTransition {
+                binding.constraintLayout.prepareTransition {
                     moveResize()
                     changeText {
-                        +collapseButton
+                        +binding.collapseButton
                         changeTextBehavior = ChangeText.CHANGE_BEHAVIOR_OUT_IN
                     }
                 }
-                collapseButton.setText(R.string.collapse)
+                binding.collapseButton.setText(R.string.collapse)
             }
             duration = 300
         }
-        expandConstraint.applyTo(constraintLayout)
-        metamorphosisDesc2.isGone = false
-        metamorphosisDesc.isGone = true
+        expandConstraint.applyTo(binding.constraintLayout)
+        binding.metamorphosisDesc2.isGone = false
+        binding.metamorphosisDesc.isGone = true
     }
 }
