@@ -2,7 +2,7 @@ package `in`.arunkumarsampath.transitionx.sample.home.transitionsamples.cart
 
 import android.os.Parcelable
 import androidx.annotation.DrawableRes
-import kotlinx.android.parcel.Parcelize
+import kotlinx.parcelize.Parcelize
 
 @Parcelize
 data class CartItem(
